Update file_permissions.rst
I have a mac and am a previous linux user. I just spent around an hour trying to get behat the permissions it needs to run tests. All other permissions were set fine with the command as it is written, but it seems recursive permissions are required for some aspect of behat. 

I am unsure how this should best be added, but at least adding a note to try recursive permissions if you have tried the other version and are still running into a wall. 

Thanks for all you do!